correct onlyoneclue.cpp solution that used size_t instead of int

+++ Only One Clue/onlyoneclue.cpp
@@ -1,13 +1,12 @@
 #include <bits/stdc++.h>
- using namespace std;
- void test(){
-     int k,x;
-     cin>>k>>x;
-     for(size_t i=x-k+1;i<x+k;i++)
-     cout<<i<<" ";
-
- }
- int main(){ 
-     test();
-     return 0;
- }
+using namespace std;
+void test() {
+    int k, x;
+    cin >> k >> x;
+    for (int i = x - k + 1; i < x + k; i++)
+        cout << i << " ";
+}
+int main() {
+    test();
+    return 0;
+}
